a thread can only be deleted by the creator

+++ app/Controllers/Http/ThreadController.js
@@ -9,9 +9,14 @@ class ThreadController {
     return response.json({ thread });
   }
 
-  async destroy({ params }) {
+  async destroy({ params, auth, response }) {
     const thread = await Thread.findOrFail(params.id);
+    if (thread.user_id !== auth.user.id) {
+      return response.forbidden();
+    }
     await thread.delete();
+
+    return response.status(204).json({ message: 'thread delete' });
   }
 }
